Update types, move type defs into audit.model.ts, user ESModule export style
